requirements.txt
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
openai>=1.0.0
watchdog>=3.0.0
pystray>=0.19.0
Pillow>=9.0.0
python-dotenv>=1.0.0
